requirements.txt

annotated-types==0.6.0
asgiref==3.7.2
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
greenlet==3.0.3
mysql-connector==2.2.9
mysqlclient==2.2.4
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic_core==2.16.2
python-dotenv==1.0.1
pytz==2024.1
SQLAlchemy==2.0.27
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
